requirements.txt
streamlit
streamlit-option-menu
google-api-python-client
python-docx
PyPDF2
pytesseract
pdf2image
google-generativeai 

llama-index
llama-parse

pandas
pydantic
python-dotenv

langchain
langchain-core
Pillow
requests
